Restructure and slightly refactor code (#13)
* separate structs out into files

* no direct access to modules

* fix a clippy error

* lots of clippy warnings

* add `into_data()`

* make Yuv fields not crate-public

* remove doc(hidden) and turn modules not-public

* rustfmt

* Allow clippy::missing_const_for_fn for into_data
See https://github.com/rust-lang/rust-clippy/issues/9271, this is a bug in clippy

* fix legitimate clippy errors

+++ src/yuv_rgb/color.rs
@@ -218,12 +218,7 @@ pub fn rgb_to_yuv<T: Pixel>(
             [res[0], res[1], res[2]]
         })
         .collect::<Vec<_>>();
-    Ok(ypbpr_to_ycbcr(
-        &yuv,
-        width as usize,
-        height as usize,
-        config,
-    ))
+    Ok(ypbpr_to_ycbcr(&yuv, width, height, config))
 }
 
 pub fn transform_primaries(
